+++ R/nhanes_sample.R
@@ -1,0 +1,12 @@
+#' @docType data
+#' @title Sample dataset from the National Health and Nutrition Examination Survey
+#' @name nhanes_sample
+#' @keywords data
+#'
+#' @description Selected variables from the National Health and Nutrition Examination
+#'              Survey that are used in the example from Lumley (2010), Appendix E.
+#'              See \code{\link{svyglm.nb}} for examples.
+#'
+#' @references Lumley T (2010). Complex Surveys: a guide to analysis using R. Wiley
+NULL
+